requirements.txt
fastapi[all]
fastapi-mail
passlib[bcrypt]
python-jose[cryptography]
pymongo[srv]
pandas
rich
openpyxl
uvicorn
